+++ silence/server/default_endpoints.py
@@ -19,7 +19,7 @@ def show_api_endpoints():
     return jsonify(server_manager.API_SUMMARY.get_endpoint_list()), 200
 
 def login():
-    USERS_TABLE, IDENTIFIER_FIELD, PASSWORD_FIELD = get_login_settings()
+    USERS_TABLE, IDENTIFIER_FIELD, PASSWORD_FIELD, ROLE_FIELD = get_login_settings()
     # Ensure that the user has sent the required fields
     form = request.json if request.is_json else request.form
     form = filter_fields_db(form, USERS_TABLE)
@@ -62,7 +62,7 @@ def login():
     return jsonify(res), 200
 
 def register():
-    USERS_TABLE, IDENTIFIER_FIELD, PASSWORD_FIELD = get_login_settings()
+    USERS_TABLE, IDENTIFIER_FIELD, PASSWORD_FIELD, ROLE_FIELD = get_login_settings()
     
     # Ensure that the user has sent the required fields
     form = request.json if request.is_json else request.form
@@ -87,6 +87,10 @@ def register():
     # Create the user object, replacing the password with the hashed one
     user = dict(form)
     user[PASSWORD_FIELD] = generate_password_hash(password)
+
+    # Assign a default role to the user, if specified in the settings
+    if settings.DEFAULT_ROLE_REGISTER:
+        user[ROLE_FIELD] = settings.DEFAULT_ROLE_REGISTER
 
     # Try to insert it in the DB
     # Since the /register endpoint must adapt to any possible table,
@@ -143,4 +147,9 @@ def get_login_settings():
     users_table = settings.USER_AUTH_DATA["table"]
     identifier_field = col_correct_case(settings.USER_AUTH_DATA["identifier"], users_table)
     password_field = col_correct_case(settings.USER_AUTH_DATA["password"], users_table)
-    return users_table, identifier_field, password_field
+    if "role" in settings.USER_AUTH_DATA:
+        role_field = col_correct_case(settings.USER_AUTH_DATA["role"], users_table)
+    else:
+        role_field = None
+
+    return users_table, identifier_field, password_field, role_field